+++ HISTORY.md
@@ -3,10 +3,13 @@
 ## in development:
 * **New Features:**
     * expose a copy of the custom block as "caller" to input slot reaction scripts inside custom block definitions
+* **Notable Changes:**
+    * automatically add a dropdown menu to any custom block input slot whose definition has a slot menu event hat block for it
 
 ### 2024-11-11
 * threads: expose a copy of the custom block as "caller" to input slot reaction scripts inside custom block definitions
 * blocks: reversed order of slot-event dropdown items
+* byob: automatically add a dropdown menu to any custom block input slot whose definition has a slot menu event hat block for it
 
 ## 10.2.2:
 * **Notable Changes:**